requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-model-utils==4.5.1
djangorestframework==3.15.1
Markdown==3.6
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
psycopg2-binary == 2.9.9
openapi==1.1.0
drf-yasg==1.21.7